SIG-425 Moved signals views to signals django app. Cleanup urls configuration.
